+++ atdgen/test/test.expected.mli
@@ -131,13 +131,13 @@ type even_natural = Test_lib.Even_natural.t
   \}\}\}abc[def]ghi
   
 {v
-j  *  j
- k * k
-  l*l
+  j  *  j
+   k * k
+    l*l
 v}
   
 {v
-mno
+  mno
 v}
   
   [pqr]\{stu\}vwx
@@ -147,7 +147,7 @@ v}
   [\} \[ \] \{v]
   
 {v
-\} [x] v\} \{v [ @ 
+  \} [x] v\} \{v [ @ 
 v}
 *)
 type def = Test_lib.Biniou.def